refactor(FontAwesome): rename library export file
